+++ lib/KnockPattern/PatternStorage/PatternStorage.h
@@ -15,9 +15,11 @@ bool initPatternStorage();
  * @brief Store a pattern in the pattern storage
  *
  * @param pattern the pattern to store
+ * @param existingPatterns the existing patterns in the storage, if NULL, the patterns will be loaded
+ * @param existingPatternsLen the length of the existing patterns
  * @return true if the pattern was stored successfully
  */
-bool storePattern(PatternData* pattern);
+bool storePattern(PatternData* pattern, PatternData* existingPatterns, size_t existingPatternsLen);
 
 /**
  * @brief Load all patterns from the pattern storage