+++ lib/KnockPattern/PatternTypes.h
@@ -4,6 +4,10 @@
 HEADER_BEGIN
 
 typedef struct {
+    int id;
+    long* deltaTimes;
+    size_t lengthPattern;
+    unsigned long durationMillis;
 } PatternData;
 
 HEADER_END