+++ lib/String/stringHelper.h
@@ -25,4 +25,45 @@ int stringToInt(const char* str, int base);
  */
 int* splitString(const char* str, const char* delim, int* len);
 
+/**
+ * @brief Get a substring of a string
+ * WARNING: This function allocates memory for the substring, make sure to free it after usage
+ *
+ * @param str the string to get the substring from
+ * @param start the start index of the substring
+ * @param end the end index of the substring
+ * @return char* the substring
+ */
+char* substring(const char* str, int start, int end);
+
+/**
+ * @brief Concatenate two strings
+ * WARNING: This function allocates memory for the concatenated string, make sure to free it after usage
+ *
+ * @param str1 the first string
+ * @param str2 the second string
+ * @return char* the concatenated string
+ */
+char* concat(const char* str1, const char* str2);
+
+/**
+ * @brief Concatenate three strings
+ * WARNING: This function allocates memory for the concatenated string, make sure to free it after usage
+ *
+ * @param str1 the first string
+ * @param str2 the second string
+ * @param str3 the third string
+ * @return char* the concatenated string
+ */
+char* concat3(const char* str1, const char* str2, const char* str3);
+
+/**
+ * @brief Generate a random string
+ * WARNING: This function allocates memory for the random string, make sure to free it after usage
+ *
+ * @param len the length of the random string
+ * @return char* the random string
+ */
+char* randomString(int len);
+
 HEADER_END